Refactor TODO.md: Add secondary language option for loading data in English if Vietnamese is not available
