+++ reelphim-app/src/components/MovieGrid.tsx
