+++ reelphim-app/src/components/MovieSection.tsx
