+++ reelphim-app/src/config/siteConfig.ts
@@ -11,7 +11,7 @@ export const SITE_CONFIG = {
     },
     CONTACT_EMAIL: 'support@reelphim.com',
     SUPPORT_PHONE: '+84 123 456 789',
-    LANGUAGE: 'vi-VN',
+    LANGUAGE: 'vi-VN', 
   };
   
   export const NAVIGATION = {
