+++ reelphim-app/src/main.jsx
@@ -1,7 +1,7 @@
 // MAIN.JSX VUI LÒNG KHÔNG CHỈNH SỬA FILE NÀY CHO CÁC TÍNH NĂNG CỦA WEB.
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './App.jsx'
+import App from './App.tsx'
 import './index.css'
 import './fonts.css'
 createRoot(document.getElementById('root')).render(
