+++ reelphim-app/src/pages/GenrePage.tsx
