+++ reelphim-app/src/pages/MovieDetailPage.tsx
